chore: move dropzone in its own package
