+++ packages/react-front-kit-dropzone/README.md
@@ -1,0 +1,19 @@
+# React Front Kit dropzone
+
+Dropzone component.
+
+## Prerequisite
+
+You must have React 18 or later installed on your project.
+
+## Installation
+
+Install `@smile/react-front-kit-dropzone` and peer dependencies with:
+
+```bash
+npm i @smile/react-front-kit-dropzone @emotion/react @mantine/core @mantine/hooks @mantine/styles @mantine/dropzone @phosphor-icons/react
+```
+
+## License
+
+MIT
