+++ packages/react-front-kit-dropzone/babel.config.js
@@ -1,0 +1,3 @@
+const { babelConfig } = require('test');
+
+module.exports = babelConfig;
