+++ packages/react-front-kit-dropzone/jest.config.ts
@@ -1,0 +1,1 @@
+export { jestConfig as default } from 'test';
