+++ packages/react-front-kit-dropzone/package.json
@@ -1,0 +1,73 @@
+{
+  "name": "@smile/react-front-kit-dropzone",
+  "version": "0.2.0",
+  "description": "Dropzone React component library based on mantine",
+  "license": "MIT",
+  "homepage": "https://github.com/Smile-SA/react-front-kit",
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:Smile-SA/react-front-kit.git"
+  },
+  "bugs": {
+    "url": "https://github.com/Smile-SA/react-front-kit/issues"
+  },
+  "keywords": [
+    "react",
+    "react-component",
+    "next",
+    "nextjs",
+    "ui",
+    "components",
+    "ui-kit",
+    "library",
+    "frontend",
+    "design",
+    "emotion",
+    "dropzone"
+  ],
+  "author": {
+    "name": "Tony Cabaye",
+    "email": "tonai59+github@gmail.com",
+    "url": "https://github.com/tonai"
+  },
+  "main": "./dist/index.js",
+  "module": "./src/index.tsx",
+  "types": "./dist/index.d.ts",
+  "scripts": {
+    "build": "tsc --project tsconfig.build.json",
+    "test": "jest",
+    "prepublishOnly": "npm run build && node ./scripts/prepublish.mjs"
+  },
+  "dependencies": {
+    "@smile/react-front-kit": "*",
+    "@smile/react-front-kit-shared": "*"
+  },
+  "devDependencies": {
+    "@babel/preset-env": "^7.22.20",
+    "@babel/preset-react": "^7.22.15",
+    "@babel/preset-typescript": "^7.23.0",
+    "@storybook/addon-actions": "^7.4.1",
+    "@storybook/jest": "^0.2.2",
+    "@storybook/preview-api": "^7.4.1",
+    "@storybook/testing-library": "^0.2.0",
+    "@testing-library/jest-dom": "^6.1.3",
+    "@testing-library/react": "^14.0.0",
+    "@types/react": "^18.2.21",
+    "@types/react-dom": "^18.2.7",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "test": "*",
+    "tsconfig": "*",
+    "typescript": "^5.2.2"
+  },
+  "peerDependencies": {
+    "@emotion/react": ">=11",
+    "@mantine/core": "6",
+    "@mantine/dropzone": "6",
+    "@mantine/hooks": "6",
+    "@mantine/styles": "6",
+    "@phosphor-icons/react": ">=2",
+    "react": ">=18.0",
+    "react-dom": ">=18.0"
+  }
+}
