+++ packages/react-front-kit-dropzone/src/index.tsx
@@ -1,0 +1,2 @@
+/* eslint-disable react-refresh/only-export-components */
+export * from './Components/Dropzone/Dropzone';
