+++ packages/react-front-kit-dropzone/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "tsconfig/react-library.json"
+}
