+++ packages/react-front-kit-table/src/index.tsx
@@ -1,0 +1,2 @@
+/* eslint-disable react-refresh/only-export-components */
+export * from './Components/Table/Table';
