+++ packages/react-front-kit/src/Components/InfoCard/Motif.tsx
