+++ packages/react-front-kit/src/index.tsx
@@ -3,13 +3,12 @@ export * from '@smile/react-front-kit-shared';
 // component exports
 export * from './Components/TruncateString/TruncateString';
 export * from './Components/BitConverter/BitConverter';
-export * from './Components/Dropzone/Dropzone';
 export * from './Components/Breadcrumbs/Breadcrumbs';
 export * from './Components/CollapseButton/CollapseButtonControlled';
 export * from './Components/CollapseButton/CollapseButton';
 export * from './Components/ConfirmModal/ConfirmModal';
 export * from './Components/DropdownButton/DropdownButton';
-export * from './Components/DropzoneCard/DropzoneCard';
+export * from './Components/InfoCard/InfoCard';
 export * from './Components/Header/Header';
 export * from './Components/HeaderSearch/HeaderSearch';
 export * from './Components/Pagination/Pagination';
